requirements.txt
streamlit
pandas
seaborn
matplotlib
openpyxl
xlrd
plotly.express
random
